benchmark/pyproject.toml
[project]
name = "llm-benchmark"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
